orangefs: don't use iget to allocate inodes
When creating new files we don't want to use iget to allocate a new
inode, instead use new_inode to allocate a guaranteed new inode, and
use insert_inode_locked4 to make sure there's not old inode on the
way out that we were racing with.  Also clean up the surrounding
code pathes a little bit.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Mike Marshall <hubcap@omnibond.com>