orangefs: add upcall_processed to orangefs_kernel_op_s
The Orangefs V3 kernel module will use upcall and downcall trailer_bufs
as a vehicle to move V3 specific opaque data back and forth. Upcall ops with
payloads in the trailer_buf need to be read twice by the client-core,
once to get the "normal" op data and once to get the trailer_buf. The
client-core can deal with this as the out-of-tree kernel module does
this too for readx/writex. upcall_processed is pretty much
equivalent to op_linger in the out-of-tree kernel module.

Signed-off-by: Mike Marshall <hubcap@omnibond.com>